New formatter to print steps catalog
A new formatter class 'StepsCatalogFormatter' is added. The formatter is
named 'steps.catalog'.

It is intended to provide a catalog of available steps to non-technical
users, so it prints the steps as they would appear in the feature file
and ommits the step functions' names and locations.

A feature file with a test case is also provided