convert an object to unicode (py2) using __unicode__ method first
use case: text(AssertionError(u"some unicode string àèìòù")) raised a UnicodeError because it naively tried to call str() on a unicode object.

Since AssertionError defines its own __unicode__ method, it is convenient to delegate Unicode conversion to it.

Old text() behaviour is maintained for objects not defining a __unicode__ method