filters: select scenarios to run with conditional expressions
This patch adds two new configuration parameters, --filter-in, and
--filter-out, which allow users to select the scenarios to run with
conditional logic. The expressions are evaluated in a limited
environment with the python eval function. If the expression cannot be
evaluated, it is counted as having been evaluated to false.

In addition to the effective tag-set for the scenario, filters have
access to the parameters provided in an Examples table, for scenarios
derived from an Outline.

Testing:
Run scenarios having an @ipv4 tag
% behave --filter-in '"ipv4" in tags'

Dont run scenarios in features for "ipv6"
% behave --filter-out '"ipv6" in feature'

Dont run scenarios named "dhcp"
% behave --filter-out 'scenario.lower() == "dhcp"'

Run scenarios with "tcp" in a column named "ipproto" in an Examples row
% behave --filter-in 'ipproto == "tcp"'

Work as expected