runner: big refactoring of "parallel" mode
This is based on the "behave-parallel" project, but now using a
different approach: the master node only scans the features and
steps, then serializes the `ids` of each job to multiprocessing
slaves. Results are also serialized as primitive dicts back to
the master. The master will then call formatters and reporters,
using the de-serialized results.