Formatters: allow steps.usage to select what to print
For testing of behave repository itself, it'd be useful to have
formatters that would report just faulty steps. The steps.usage
formatter now read 'steps.usage.classes' user config key and if it's
content is comma-separated list of (used|unused|undefined). If it is not
recognized this way, all the step classes are printed to maintain the
present behaviour for user who could already depend on this very key.

Example usage:
$ behave --dry -f steps.usage -D steps.usage.classes=unused,undefined